requirements.txt
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.5
lxml==4.9.3
python-dotenv==1.0.0
Pillow==10.0.1
validators==0.22.0
